src/requirements.txt
Flask==2.2.3
pyyaml==5.3.1
apscheduler==3.6.3
